Improve salt_lsp: deal with jinja
It's kind of a spraling change but the core of the change is to change
the parsing of salt's state file.
Implemented a naive way to resolve the Jinja code so that the generated
file is a valid yaml file.

Also deal with the fact that poorly aligned yaml file are valid in Salt,
for instance:

```
foo:
  file.managed:
  - source: salt://myfile
  - require:
    - file: bazr
```